requirements.txt
Django
django-allauth
CKEditor 5